refactor: replace async_trait implementations with macro for pagination in client modules
