+++ notionrs/src/client/block/get_block_children.rs
@@ -10,28 +10,10 @@ pub struct GetBlockChildrenClient {
     pub(crate) start_cursor: Option<String>,
 }
 
-#[async_trait::async_trait]
-impl crate::r#trait::Paginate<notionrs_types::object::block::BlockResponse>
-    for GetBlockChildrenClient
-{
-    fn paginate_start_cursor(self, start_cursor: Option<String>) -> Self {
-        match start_cursor {
-            Some(c) => self.start_cursor(c),
-            None => self,
-        }
-    }
-
-    async fn paginate_send(
-        self,
-    ) -> Result<
-        notionrs_types::object::response::ListResponse<
-            notionrs_types::object::block::BlockResponse,
-        >,
-        crate::error::Error,
-    > {
-        Ok(self.send().await?)
-    }
-}
+crate::impl_paginate!(
+    GetBlockChildrenClient,
+    notionrs_types::object::block::BlockResponse
+);
 
 impl Default for GetBlockChildrenClient {
     fn default() -> Self {
