+++ notionrs/src/client/comment/retrieve_comments.rs
@@ -12,24 +12,10 @@ pub struct RetrieveCommentsClient {
     pub(crate) start_cursor: Option<String>,
 }
 
-#[async_trait::async_trait]
-impl crate::r#trait::Paginate<notionrs_types::object::comment::Comment> for RetrieveCommentsClient {
-    fn paginate_start_cursor(self, start_cursor: Option<String>) -> Self {
-        match start_cursor {
-            Some(c) => self.start_cursor(c),
-            None => self,
-        }
-    }
-
-    async fn paginate_send(
-        self,
-    ) -> Result<
-        notionrs_types::object::response::ListResponse<notionrs_types::object::comment::Comment>,
-        crate::error::Error,
-    > {
-        Ok(self.send().await?)
-    }
-}
+crate::impl_paginate!(
+    RetrieveCommentsClient,
+    notionrs_types::object::comment::Comment
+);
 
 #[derive(Serialize)]
 struct RetrieveCommentsQueryParams {
