+++ notionrs/src/client/search/search_database.rs
@@ -19,28 +19,10 @@ pub struct SearchDatabaseClient {
     pub(crate) page_size: Option<u32>,
 }
 
-#[async_trait::async_trait]
-impl crate::r#trait::Paginate<notionrs_types::object::database::DatabaseResponse>
-    for SearchDatabaseClient
-{
-    fn paginate_start_cursor(self, start_cursor: Option<String>) -> Self {
-        match start_cursor {
-            Some(c) => self.start_cursor(c),
-            None => self,
-        }
-    }
-
-    async fn paginate_send(
-        self,
-    ) -> Result<
-        notionrs_types::object::response::ListResponse<
-            notionrs_types::object::database::DatabaseResponse,
-        >,
-        crate::error::Error,
-    > {
-        Ok(self.send().await?)
-    }
-}
+crate::impl_paginate!(
+    SearchDatabaseClient,
+    notionrs_types::object::database::DatabaseResponse
+);
 
 #[derive(Debug, Serialize, Deserialize, Default)]
 pub struct SearchDatabaseRequestBody {
