+++ notionrs/src/lib.rs
@@ -3,6 +3,7 @@
 
 pub mod client;
 pub mod error;
+pub(crate) mod r#macro;
 pub mod r#trait;
 
 pub use crate::client::Client;
